+++ packages/react-html/src/common.ts
@@ -1,5 +1,6 @@
 export * from './components';
-export {default as Manager, EFFECT_ID} from './manager';
-export {Provider} from './context';
+export {HtmlManager, EFFECT_ID} from './manager';
+export {HtmlContext} from './context';
 export {showPage, getSerialized} from './utilities';
-export {createSerializer} from './serializer';
+export {useDomEffect} from './hooks';
+export {createSerializer, useSerialized} from './serializer';