+++ packages/react-html/src/manager.ts
@@ -17,7 +17,7 @@ interface Subscription {
 
 export const EFFECT_ID = Symbol('html');
 
-export default class Manager {
+export class HtmlManager {
   effect: EffectKind = {
     id: EFFECT_ID,
     betweenEachPass: () => this.reset(),
@@ -57,6 +57,9 @@ export default class Manager {
 
   subscribe(subscription: Subscription) {
     this.subscriptions.add(subscription);
+    return () => {
+      this.subscriptions.delete(subscription);
+    };
   }
 
   addTitle(title: string) {