+++ packages/react-network/src/components/ConnectSource.tsx
@@ -1,13 +1,11 @@
-import * as React from 'react';
 import {CspDirective} from '@shopify/network';
-import SetCspDirective from './SetCspDirective';
+import {useCspDirective} from '../hooks';
 
 interface Props {
   sources: string | string[];
 }
 
 export default function ConnectSource({sources}: Props) {
-  return (
-    <SetCspDirective directive={CspDirective.ConnectSrc} value={sources} />
-  );
+  useCspDirective(CspDirective.ConnectSrc, sources);
+  return null;
 }