+++ packages/react-network/src/components/ManifestSource.tsx
@@ -1,13 +1,11 @@
-import * as React from 'react';
 import {CspDirective} from '@shopify/network';
-import SetCspDirective from './SetCspDirective';
+import {useCspDirective} from '../hooks';
 
 interface Props {
   sources: string | string[];
 }
 
 export default function ManifestSource({sources}: Props) {
-  return (
-    <SetCspDirective directive={CspDirective.ManifestSrc} value={sources} />
-  );
+  useCspDirective(CspDirective.ManifestSrc, sources);
+  return null;
 }