+++ packages/react-network/src/components/UpgradeInsecureRequests.tsx
@@ -1,16 +1,11 @@
-import * as React from 'react';
 import {CspDirective} from '@shopify/network';
-import SetCspDirective from './SetCspDirective';
+import {useCspDirective} from '../hooks';
 
 interface Props {
   value?: boolean;
 }
 
 export default function UpgradeInsecureRequests({value = true}: Props) {
-  return (
-    <SetCspDirective
-      directive={CspDirective.UpgradeInsecureRequests}
-      value={value}
-    />
-  );
+  useCspDirective(CspDirective.UpgradeInsecureRequests, value);
+  return null;
 }