+++ packages/react-network/src/components/WebRtcSource.tsx
@@ -1,11 +1,11 @@
-import * as React from 'react';
 import {CspDirective} from '@shopify/network';
-import SetCspDirective from './SetCspDirective';
+import {useCspDirective} from '../hooks';
 
 interface Props {
   sources: string | string[];
 }
 
 export default function WebRtcSource({sources}: Props) {
-  return <SetCspDirective directive={CspDirective.WebrtcSrc} value={sources} />;
+  useCspDirective(CspDirective.WebrtcSrc, sources);
+  return null;
 }